backend/requirements.txt
flask>=3.0,<4.0
flask-cors>=4.0,<5.0
sqlalchemy>=2.0.30,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1.18,<4.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
pytest>=7.4,<9.0
